Formatting to assertJ style as best I can.
I tried importing the Elipse settings to intellij, but they reformatted
existing code (either the imported settings didn't work, or the existing
code did not conform). Either way, i've formatted the new code by code
(as best I can) to match the style of the exising code. (in the assertJ
guava OptionalAssert class and it's tests).